cleanup and make similar popover a pure component, wrap logic in hooks
