+++ app/packages/core/src/components/Actions/Similarity/constants.ts
@@ -1,3 +1,4 @@
+// Duplicated from @fiftyone/similarity-search/constants to avoid
+// a cross-package dependency (core does not depend on similarity-search)
 export const SEARCH_OPERATOR_URI = "@voxel51/panels/similarity_search";
-export const INIT_RUN_OPERATOR_URI = "@voxel51/panels/init_similarity_run";
 export const PANEL_NAME = "similarity_search_panel";
